requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
torch>=2.1
torchvision>=0.16
transformers>=4.36
Pillow>=10.0
omegaconf>=2.3
